Proyecto 2 - Programación y algoritmos: *Código terminado. *Falta refinamiento en el Makefile.

+++ Semestre 1/Programacion y algoritmos/prog2017_Proyecto2_Alvarez_ES/individual.cpp
@@ -1,10 +1,10 @@
 #include "individual.hpp"
 
 Individual::Individual() {
-	this->quality = 0;
+	this->quality = 0.0;
 }
 
-Individual::Individual(vector<int> bv, int quality) {
+Individual::Individual(vector<int> bv, double quality) {
 	this->bv = bv;
 	this->quality = quality;
 }
@@ -15,7 +15,7 @@ void Individual::set_bv(vector<int> bv) {
 	this->bv = bv;
 }
 
-void Individual::set_quality(int quality) {
+void Individual::set_quality(double quality) {
 	this->quality = quality;
 }
 
@@ -23,6 +23,6 @@ vector<int> Individual::get_bv() {
 	return bv;
 }
 
-int Individual::get_quality() const {
+double Individual::get_quality() const {
 	return quality;
 }
